requirements.txt
numpy
torchaudio
scikit-learn
pandas
rtpt
tqdm
torchinfo
black
